Support multithreading in two steps: a) buildSortedReadEndList, and b) the final makr duplicate stage.
Multithreading is implemented as one producer thread and one consumer thread, each
processes one SAMrecord at a time. Producer thread communicates with consumer thread
through a circular buffer.

For efficency, the circular buffer is implemented in a lock-free fashion. It does not
support multi-producer or multi-consumer case.

Note that this buffer does not have a way to detect if it is being accessed by
multiple producers or multiple consumers. Use it cautiously.